+++ src/components/config-editor/config-editor.test.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { shallow, ShallowWrapper } from 'enzyme';
 import { DataSourceSettings } from '@grafana/data';
 import { RadioButtonGroup } from '@grafana/ui';
-import { ClientTypeValue, RedisDataSourceOptions } from '../../types';
+import { ClientTypeValue } from '../../constants';
+import { RedisDataSourceOptions } from '../../types';
 import { ConfigEditor } from './config-editor';
 
 /**
