+++ src/types.ts
@@ -1,4 +1,5 @@
 import { DataSourceJsonData } from '@grafana/data';
+import { ClientTypeValue } from './constants';
 
 /**
  * Options configured for each DataSource instance
@@ -107,23 +108,3 @@ export interface RedisSecureJsonData {
    */
   tlsCACert?: string;
 }
-
-/**
- * Client Type Values
- */
-export enum ClientTypeValue {
-  CLUSTER = 'cluster',
-  SENTINEL = 'sentinel',
-  SOCKET = 'socket',
-  STANDALONE = 'standalone',
-}
-
-/**
- * Client Types
- */
-export const ClientType = [
-  { label: 'Standalone', value: ClientTypeValue.STANDALONE },
-  { label: 'Cluster', value: ClientTypeValue.CLUSTER },
-  { label: 'Sentinel', value: ClientTypeValue.SENTINEL },
-  { label: 'Socket', value: ClientTypeValue.SOCKET },
-];
